+++ example/data_source_virtual_environment_cluster_alias.tf
@@ -4,6 +4,6 @@ data "proxmox_virtual_environment_cluster_alias" "example" {
   name = proxmox_virtual_environment_cluster_alias.example.name
 }
 
-output "data_proxmox_virtual_environment_cluster_alias_example_cidr" {
-  value = proxmox_virtual_environment_cluster_alias.example.cidr
+output "data_proxmox_virtual_environment_cluster_alias_example" {
+  value = proxmox_virtual_environment_cluster_alias.example
 }