+++ example/data_source_virtual_environment_group.tf
@@ -2,14 +2,6 @@ data "proxmox_virtual_environment_group" "example" {
   group_id = proxmox_virtual_environment_group.example.id
 }
 
-output "data_proxmox_virtual_environment_group_example_acl" {
-  value = data.proxmox_virtual_environment_group.example.acl
-}
-
-output "data_proxmox_virtual_environment_group_example_comment" {
-  value = data.proxmox_virtual_environment_group.example.comment
-}
-
-output "data_proxmox_virtual_environment_group_example_members" {
-  value = data.proxmox_virtual_environment_group.example.members
+output "data_proxmox_virtual_environment_group_example" {
+  value = data.proxmox_virtual_environment_group.example
 }