+++ example/data_source_virtual_environment_roles.tf
@@ -2,14 +2,6 @@ data "proxmox_virtual_environment_roles" "example" {
   depends_on = [proxmox_virtual_environment_role.example]
 }
 
-output "data_proxmox_virtual_environment_roles_example_privileges" {
-  value = data.proxmox_virtual_environment_roles.example.privileges
-}
-
-output "data_proxmox_virtual_environment_roles_example_role_ids" {
-  value = data.proxmox_virtual_environment_roles.example.role_ids
-}
-
-output "data_proxmox_virtual_environment_roles_example_special" {
-  value = data.proxmox_virtual_environment_roles.example.special
+output "data_proxmox_virtual_environment_roles_example" {
+  value = data.proxmox_virtual_environment_roles.example
 }