+++ proxmox/virtual_environment_users_types.go
@@ -31,32 +31,53 @@ type VirtualEnvironmentUserGetResponseBody struct {
 
 // VirtualEnvironmentUserGetResponseData contains the data from an user get response.
 type VirtualEnvironmentUserGetResponseData struct {
-	Comment        *string          `json:"comment,omitempty"`
-	Email          *string          `json:"email,omitempty"`
-	Enabled        *CustomBool      `json:"enable,omitempty"`
-	ExpirationDate *CustomTimestamp `json:"expire,omitempty"`
-	FirstName      *string          `json:"firstname,omitempty"`
-	Groups         *[]string        `json:"groups,omitempty"`
-	Keys           *string          `json:"keys,omitempty"`
-	LastName       *string          `json:"lastname,omitempty"`
+	Comment        *string                                                `json:"comment,omitempty"`
+	Email          *string                                                `json:"email,omitempty"`
+	Enabled        *CustomBool                                            `json:"enable,omitempty"`
+	ExpirationDate *CustomTimestamp                                       `json:"expire,omitempty"`
+	FirstName      *string                                                `json:"firstname,omitempty"`
+	Groups         *[]string                                              `json:"groups,omitempty"`
+	Keys           *string                                                `json:"keys,omitempty"`
+	LastName       *string                                                `json:"lastname,omitempty"`
+	Tokens         *map[string]VirtualEnvironmentUserTokenGetResponseData `json:"tokens,omitempty"`
 }
 
-// VirtualEnvironmentUserListResponseBody contains the body from an user list response.
+// VirtualEnvironmentUserListRequestBody contains the data for a user list request.
+type VirtualEnvironmentUserListRequestBody struct {
+	Enabled *CustomBool `json:"enabled,omitempty" url:"enabled,omitempty,int"`
+	Full    *CustomBool `json:"full,omitempty" url:"full,omitempty,int"`
+}
+
+// VirtualEnvironmentUserListResponseBody contains the body from a user list response.
 type VirtualEnvironmentUserListResponseBody struct {
 	Data []*VirtualEnvironmentUserListResponseData `json:"data,omitempty"`
 }
 
-// VirtualEnvironmentUserListResponseData contains the data from an user list response.
+// VirtualEnvironmentUserListResponseData contains the data from a user list response.
 type VirtualEnvironmentUserListResponseData struct {
-	Comment        *string          `json:"comment,omitempty"`
-	Email          *string          `json:"email,omitempty"`
-	Enabled        *CustomBool      `json:"enable,omitempty"`
-	ExpirationDate *CustomTimestamp `json:"expire,omitempty"`
-	FirstName      *string          `json:"firstname,omitempty"`
-	Groups         *[]string        `json:"groups,omitempty"`
-	ID             string           `json:"userid"`
-	Keys           *string          `json:"keys,omitempty"`
-	LastName       *string          `json:"lastname,omitempty"`
+	Comment        *string                                       `json:"comment,omitempty"`
+	Email          *string                                       `json:"email,omitempty"`
+	Enabled        *CustomBool                                   `json:"enable,omitempty"`
+	ExpirationDate *CustomTimestamp                              `json:"expire,omitempty"`
+	FirstName      *string                                       `json:"firstname,omitempty"`
+	Groups         *CustomCommaSeparatedList                     `json:"groups,omitempty"`
+	ID             string                                        `json:"userid"`
+	Keys           *string                                       `json:"keys,omitempty"`
+	LastName       *string                                       `json:"lastname,omitempty"`
+	Tokens         *[]VirtualEnvironmentUserTokenGetResponseData `json:"tokens,omitempty"`
+}
+
+// VirtualEnvironmentUserTokenGetResponseBody contains the body from a user token get response.
+type VirtualEnvironmentUserTokenGetResponseBody struct {
+	Data *VirtualEnvironmentUserTokenGetResponseData `json:"data,omitempty"`
+}
+
+// VirtualEnvironmentUserTokenGetResponseData contains the data from a user token get response.
+type VirtualEnvironmentUserTokenGetResponseData struct {
+	Comment             *string          `json:"comment,omitempty"`
+	ExpirationDate      *CustomTimestamp `json:"expire,omitempty"`
+	ID                  *string          `json:"tokenid,omitempty"`
+	PrivilegeSeperation *CustomBool      `json:"privsep,omitempty"`
 }
 
 // VirtualEnvironmentUserUpdateRequestBody contains the data for an user update request.